Reduce initial UI lag when typing in the searchbar
Unlike MiniSearch, microfuzz returns the entire list when the query text
is an empty string. This causes a significant lag when typing the first
letter in the search bar, since the unfiltered list will be rendered
first, and then rerender with the filtered list that matches the typed
letter/s.

To solve this, the `Autosuggest` component is passed an empty array for
the `suggestions` prop if `queryText` is an empty string, otherwise it
uses the actual `searchResults` returned by `useFuzzySearchList`.

I also moved most of the state inside the `GameSearh` component to avoid
passing too many props, especially since we're using Redux anyway.

Work needs to be done to fix a bug when using the up and down arrow keys
to navigate through the list of game results.